Rename dynamic quick start card to QuickStartDynamicCard
